requirements.txt
numpy
panda
scipy
matplotlib
loguru
gisplit @ git+https://github.com/jararias/gisplit@main
